+++ src/com/yanncebron/m68kplugin/lang/psi/M68kMnemonic.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Authors
+ * Copyright 2026 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -18,8 +18,10 @@ package com.yanncebron.m68kplugin.lang.psi;
 
 import com.intellij.psi.tree.IElementType;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.TestOnly;
 
 import java.util.Set;
+import java.util.function.Function;
 
 /**
  * @see M68kMnemonicRegistry
@@ -28,18 +30,12 @@ public record M68kMnemonic(IElementType elementType,
                            Set<M68kDataSize> dataSizes,
                            M68kOperand sourceOperand,
                            M68kOperand destinationOperand,
-                           Set<M68kCpu> cpus) {
+                           Set<M68kCpu> cpus,
+                           PrivilegedType privilegedType) {
 
+  @TestOnly
   public M68kMnemonic(IElementType elementType, Set<M68kDataSize> dataSizes, M68kOperand sourceOperand, M68kOperand destinationOperand) {
-    this(elementType, dataSizes, sourceOperand, destinationOperand, M68kCpu.GROUP_68000_UP);
-  }
-
-  public M68kMnemonic(IElementType elementType, Set<M68kDataSize> dataSizes, M68kOperand sourceOperand) {
-    this(elementType, dataSizes, sourceOperand, M68kCpu.GROUP_68000_UP);
-  }
-
-  public M68kMnemonic(IElementType elementType, Set<M68kDataSize> dataSizes, M68kOperand sourceOperand, Set<M68kCpu> cpus) {
-    this(elementType, dataSizes, sourceOperand, M68kOperand.NONE, cpus);
+    this(elementType, dataSizes, sourceOperand, destinationOperand, M68kCpu.GROUP_68000_UP, PrivilegedType.NONE);
   }
 
   public boolean isDeprecated() {
@@ -47,6 +43,33 @@ public record M68kMnemonic(IElementType elementType,
 
     return (sourceOperand() == M68kOperand.DATA && destinationOperand() == M68kOperand.ALTERABLE_DATA) ||
       (sourceOperand() == M68kOperand.ADDRESS_REGISTER && destinationOperand() == M68kOperand.ALTERABLE);
+  }
+
+  public enum PrivilegedType {
+    /**
+     * Never privileged.
+     */
+    NONE(m68kCpu -> Boolean.FALSE),
+
+    /**
+     * Always privileged.
+     */
+    PRIVILEGED(m68kCpu -> Boolean.TRUE),
+
+    /**
+     * Privileged for 68010 or above only.
+     */
+    PRIVILEGED_68010_ABOVE(M68kCpu.GROUP_68010_UP::contains);
+
+    private final Function<M68kCpu, Boolean> privilegedFunction;
+
+    PrivilegedType(Function<M68kCpu, Boolean> privilegedFunction) {
+      this.privilegedFunction = privilegedFunction;
+    }
+
+    public boolean isPrivileged(M68kCpu m68kCpu) {
+      return privilegedFunction.apply(m68kCpu);
+    }
   }
 
   @Override
@@ -64,6 +87,7 @@ public record M68kMnemonic(IElementType elementType,
       ", dst=" + destinationOperand +
       ", " + dataSizes +
       ", " + cpuText +
+      (privilegedType != PrivilegedType.NONE ? ", " + privilegedType.name() : "") +
       '}';
   }
 }