+++ tests/com/yanncebron/m68kplugin/lang/highlighting/M68kSyntaxAnnotatorTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 The Authors
+ * Copyright 2026 The Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,6 +30,7 @@ public class M68kSyntaxAnnotatorTest extends BasePlatformTestCase {
          <info descr="Privileged instruction" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">or #2,SR</info>
          <info descr="Privileged instruction" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">ori #2,SR</info>
          <info descr="Privileged instruction" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">eori #2,SR</info>
+         <info descr="Privileged instruction (MC68010+)" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">move.w SR,d6</info>
          <info descr="Privileged instruction" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">move.l usp,a0</info>
          <info descr="Privileged instruction" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">move.w d0,sr</info>
          <info descr="Privileged instruction" textAttributesKey="M68K_PRIVILEGED_INSTRUCTION">movec.l d0,VBR</info>